refactor: rename beagley package to beagleyai
