--- beagley/board_nonlinux.go
@@ -1,2 +1,0 @@
-// Package beagley implements a BeagleY-AI based board.
-package beagley
