+++ beagleyai/board_nonlinux.go
@@ -1,0 +1,2 @@
+// Package beagleyai implements a BeagleY-AI based board.
+package beagleyai
