+++ main.go
@@ -5,7 +5,7 @@ import (
 	"context"
 
 	"github.com/viam-modules/beagleboard/beaglebone"
-	"github.com/viam-modules/beagleboard/beagley"
+	beagleYAi "github.com/viam-modules/beagleboard/beagleyai"
 	"go.viam.com/rdk/components/board"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/module"
@@ -26,7 +26,7 @@ func mainWithArgs(ctx context.Context, args []string, logger logging.Logger) err
 		return err
 	}
 
-	if err = module.AddModelFromRegistry(ctx, board.API, beagley.Model); err != nil {
+	if err = module.AddModelFromRegistry(ctx, board.API, beagleYAi.Model); err != nil {
 		return err
 	}
